% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_formatting.R
\name{axis_width}
\alias{axis_width}
\title{Calculate a best guess \code{nameGap} value for spacing the ECharts axis titles beyond the axis labels}
\usage{
axis_width(e, e_ctxt)

axis_width(e, e_ctxt)
}
\arguments{
\item{e}{\code{echart}}

\item{e_ctxt}{\code{(character)} The context in which the e_chart is being rendered}
}
\value{
\code{num}

\code{num}
}
\description{
Calculate a best guess \code{nameGap} value for spacing the ECharts axis titles beyond the axis labels

Calculate a best guess \code{nameGap} value for spacing the ECharts axis titles beyond the axis labels
}
\seealso{
e_y_axis_formatting

e_y_axis_formatting
}
