% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{e_default_opts}
\alias{e_default_opts}
\title{Use the default options applied to each echart.}
\usage{
e_default_opts(
  e,
  dims = TRUE,
  datazoom = TRUE,
  toolbox = list(orient = "vertical", itemSize = 8, itemGap = 3),
  toolbox_feature = list(feature = c("saveAsImage", "dataZoom", "restore")),
  x_axis_formatting = TRUE,
  y_axis_formatting = TRUE,
  tooltip = list(trigger = "axis"),
  grid = list(containLabel = TRUE),
  text_style = list(fontFamily = "rennerbook", color = "#ffffff"),
  saveAsImage_filename = list(),
  ...
)
}
\arguments{
\item{dims}{\code{(list/logical/NULL)} See \link[echarts4r]{e_dims}}

\item{datazoom}{\code{(list/logical/NULL)} See \link[echarts4r]{e_datazoom}}

\item{toolbox}{\code{(list/logical/NULL)} See \link[echarts4r]{e_toolbox}}

\item{toolbox_feature}{\code{(list/logical/NULL)} See \link[echarts4r]{e_toolbox_feature}}

\item{x_axis_formatting}{\code{(list/logical/NULL)} See \code{e_x_axis_formatting}}

\item{y_axis_formatting}{\code{(list/logical/NULL)} See \code{e_y_axis_formatting}}

\item{tooltip}{\code{(list/logical/NULL)} See \link[echarts4r]{e_tooltip}}

\item{grid}{\code{(list/logical/NULL)} See \link[echarts4r]{e_grid}}

\item{saveAsImage_filename}{\code{(list/logical/NULL)} See \code{saveAsImage_filename}}

\item{...}{Additional \code{echarts4r} named arguments, IE for \link[echarts4r]{e_add_nested} used \verb{add_nested = list([...])} where the list holds the args to pass to the function.}
}
\value{
\code{echart}
}
\description{
Use the default options applied to each echart.
}
