% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_general.R, R/options.R
\name{e_get_opts}
\alias{e_get_opts}
\title{Get the options from an echart}
\usage{
e_get_opts(e)

e_get_opts(e)
}
\arguments{
\item{e}{\code{echart}}
}
\value{
\code{list}

\code{list}
}
\description{
Get the options from an echart

Get the options from an echart
}
