% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{e_has_base_opts}
\alias{e_has_base_opts}
\title{Is the echart using baseOptions}
\usage{
e_has_base_opts(e)
}
\arguments{
\item{e}{\code{echart}}
}
\value{
\code{lgl}
}
\description{
Is the echart using baseOptions
}
