% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_general.R, R/options.R
\name{e_has_group}
\alias{e_has_group}
\title{Does the echart have a group}
\usage{
e_has_group(e)

e_has_group(e)
}
\arguments{
\item{e}{\code{echart}}
}
\value{
\code{lgl}

\code{lgl}
}
\description{
Does the echart have a group

Does the echart have a group
}
