% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_general.R, R/options.R
\name{e_has_timeline}
\alias{e_has_timeline}
\title{Does the echart have a timeline}
\usage{
e_has_timeline(e)

e_has_timeline(e)
}
\arguments{
\item{e}{\code{echart}}
}
\value{
\code{lgl}

\code{lgl}
}
\description{
Does the echart have a timeline

Does the echart have a timeline
}
