% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_general.R
\name{e_is_parallel}
\alias{e_is_parallel}
\title{Is the chart a parallelAxis chart?}
\usage{
e_is_parallel(e)
}
\arguments{
\item{e}{\code{echart/htmlwidget}}
}
\value{
\code{lgl} whether or not the Echart is a parallel type chart
}
\description{
Is the chart a parallelAxis chart?
}
