% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_helpers.R
\name{e_js_axisPointer}
\alias{e_js_axisPointer}
\title{A callback function with associated dependencies for formatting Echarts axisPointers}
\usage{
e_js_axisPointer(sf = 0)
}
\arguments{
\item{sf}{\code{int} significant figures of output \strong{Default: 2}}
}
\value{
\code{shiny.tag.list}
}
\description{
A callback function with associated dependencies for formatting Echarts axisPointers
}
