% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_helpers.R
\name{e_js_num2str}
\alias{e_js_num2str}
\title{An JS callback formatting function}
\usage{
e_js_num2str(
  js_parameters = c("value", "index"),
  n = "value",
  sf = NULL,
  add_suffix = NULL,
  suffix_lb = NULL,
  format = NULL,
  magnitude = NULL
)
}
\arguments{
\item{js_parameters}{\code{chr} the parameters to pass to the Javascript Callback function}

\item{n}{\code{chr} The javascript value to pass as \code{n}, the number to be transformed by \code{num2str}}

\item{sf}{\code{int} significant figures of output}

\item{add_suffix}{\code{lgl}  whether to add a letter suffix}

\item{suffix_lb}{\code{chr} The lower bound of the divisor (IE "M" do not go lower than millions)}

\item{format}{\code{lgl} whether to reduce the magnitude when formatting}

\item{magnitude}{\code{chr} If the magnitude is known apriori this can be used to override the factoring}
}
\value{
The JS string with appropriate arguments
}
\description{
An JS callback formatting function
}
