% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_axes.R
\name{e_opts_axisPointer}
\alias{e_opts_axisPointer}
\title{Default formatter for axisPointer}
\usage{
e_opts_axisPointer(
  ...,
  show = TRUE,
  sf = 0,
  add_commas = TRUE,
  add_suffix = FALSE,
  format = TRUE,
  suffix_lb = NULL,
  magnitude = FALSE
)
}
\arguments{
\item{...}{\code{params} arbitrary Echarts axisPointer Options}

\item{show}{\code{lgl} whether to show axisPointers}

\item{sf}{\code{int} significant figures of output \strong{Default: 2}}

\item{add_commas}{\code{lgl} Format the numeric string output with commas every thousands place. \strong{Default: FALSE}}

\item{add_suffix}{\code{lgl}  whether to add a letter suffix \strong{Default: FALSE}}

\item{format}{\code{lgl} whether to reduce the magnitude when formatting \strong{Default: TRUE}}

\item{suffix_lb}{\code{chr} The lower bound of the divisor (IE "M" do not go lower than millions) \strong{Default: ''} for no lower bound}

\item{magnitude}{\code{chr} If the magnitude is known apriori this can be used to override the factoring \strong{Default: NULL}}
}
\value{
\code{list} with options and formatter function
}
\description{
Default formatter for axisPointer
}
