% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_general.R
\name{e_opts_init}
\alias{e_opts_init}
\title{Initialize the options if they don't exist}
\usage{
e_opts_init(e, ...)
}
\arguments{
\item{e}{\code{echart}}

\item{names}{\code{chr} Name of the options to initialize. All top level headings on the \link{https://echarts.apache.org/en/option.html}{ECharts Option Documentation} are viable.}
}
\value{
echart
}
\description{
Initialize the options if they don't exist
}
