% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_line.R
\name{e_opts_line_default}
\alias{e_opts_line_default}
\title{Apply default options for a}
\usage{
e_opts_line_default(
  .data,
  grid = list(left = "20\%"),
  color = NULL,
  tooltip = list(trigger = "axis", axisPointer = list(type = "cross", label =
    list(backgroundColor = "#6a7985"))),
  legend = list(textStyle = list(fontSize = 10), type = "scroll", bottom = "0\%", opacity
    = 1, padding = list(0, 0, 0, 0), itemGap = 2),
  xAxis = list(type = "category"),
  yAxis = list(type = "value"),
  ...
)
}
\arguments{
\item{.data}{\code{tbl} of Reservoir Data}

\item{grid}{See \href{https://echarts.apache.org/en/option.html#grid}{echarts setOption docs}}

\item{color}{See \href{https://echarts.apache.org/en/option.html#color}{echarts setOption docs}}

\item{tooltip}{See \href{https://echarts.apache.org/en/option.html#tooltip}{echarts setOption docs}}

\item{legend}{See \href{https://echarts.apache.org/en/option.html#legend}{echarts setOption docs}}

\item{xAxis}{See \href{https://echarts.apache.org/en/option.html#xAxis}{echarts setOption docs}}

\item{yAxis}{See \href{https://echarts.apache.org/en/option.html#yAxis}{echarts setOption docs}}

\item{...}{\code{args} Additional parameters}
}
\value{
\code{list}
}
\description{
Apply default options for a
}
