% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_line.R
\name{e_opts_make_stacked_area_series}
\alias{e_opts_make_stacked_area_series}
\title{Makes an appropriately styled data list for stacked area charts for
a single serie}
\usage{
e_opts_make_stacked_area_series(
  data,
  name = NULL,
  type = "line",
  stack = "Total",
  smooth = TRUE,
  lineStyle = list(width = 0),
  showSymbol = FALSE,
  areaStyle = list(opacity = 0.8),
  emphasis = list(focus = "series"),
  color = NULL,
  ...
)
}
\arguments{
\item{data}{\code{num} series data}

\item{name}{\code{chr} See \href{https://echarts.apache.org/en/option.html#series-line.name}{echarts setOption docs}}

\item{type}{\code{chr} See \href{https://echarts.apache.org/en/option.html#series-line.type}{echarts setOption docs}}

\item{stack}{\code{chr} See \href{https://echarts.apache.org/en/option.html#series-line.stack}{echarts setOption docs}}

\item{smooth}{\code{chr} See \href{https://echarts.apache.org/en/option.html#series-line.smooth}{echarts setOption docs}}

\item{lineStyle}{\code{chr} See \href{https://echarts.apache.org/en/option.html#series-line.lineStyle}{echarts setOption docs}}

\item{showSymbol}{\code{chr} See \href{https://echarts.apache.org/en/option.html#series-line.showSymbol}{echarts setOption docs}}

\item{areaStyle}{\code{chr} See \href{https://echarts.apache.org/en/option.html#series-line.areaStyle}{echarts setOption docs}}

\item{emphasis}{\code{chr} See \href{https://echarts.apache.org/en/option.html#series-line.emphasis}{echarts setOption docs}}

\item{color}{\code{chr} See \href{https://echarts.apache.org/en/option.html#series-line.color}{echarts setOption docs}}

\item{...}{additional parameters passed to \href{https://echarts.apache.org/en/option.html#series-line}{series}}
}
\value{
appropriately formatted list to add as an entry to the series list
in the base list
}
\description{
Makes an appropriately styled data list for stacked area charts for
a single serie
}
