% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_general.R
\name{e_opts_parallelAxisDefault}
\alias{e_opts_parallelAxisDefault}
\title{Add default options for Parallel Axis}
\usage{
e_opts_parallelAxisDefault(e, nameLocation = NULL, nameGap = NULL, ...)
}
\arguments{
\item{nameLocation}{\code{chr} \href{https://echarts.apache.org/en/option.html#parallel.parallelAxisDefault.nameLocation}{Link}}

\item{nameGap}{\code{num} \href{https://echarts.apache.org/en/option.html#parallel.parallelAxisDefault.nameGap}{Link}}

\item{axisLabel}{\code{list} \href{https://echarts.apache.org/en/option.html#parallel.parallelAxisDefault.axisLabel}{Link}}
}
\value{
\code{echart}
}
\description{
Add default options for Parallel Axis
}
\seealso{
Other options: 
\code{\link{e_default_opts}()},
\code{\link{e_get_opts}()},
\code{\link{e_has_base_opts}()},
\code{\link{e_has_group}()},
\code{\link{e_has_timeline}()},
\code{\link{e_opts_is_xAxis_series}()},
\code{\link{e_opts_line_default}()},
\code{\link{e_opts_make_stacked_area_series}()},
\code{\link{e_opts_modify}()},
\code{\link{e_opts_parallel}()},
\code{\link{e_opts_update}()},
\code{\link{e_opts_xAxis_type}()},
\code{\link{e_opts_xAxis}()},
\code{\link{e_opts_yAxis}()},
\code{\link{e_parallel_default_opts}()},
\code{\link{e_series_axis_data}()},
\code{\link{e_series_data}()},
\code{\link{e_series}()},
\code{\link{e_x_axis_formatting}()},
\code{\link{e_yAxis_width}()},
\code{\link{e_y_axis_formatting}()}
}
\concept{options}
