% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_general.R
\name{e_opts_update}
\alias{e_opts_update}
\title{Update echarts options}
\usage{
e_opts_update(e, opts, replace = FALSE)
}
\arguments{
\item{e}{\code{echart}}

\item{opts}{\code{list} of options to replace/modify}

\item{replace}{\code{lgl} whether to replace or modify (the default)}
}
\value{
\code{echarts}
}
\description{
Update echarts options
}
