% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_axes.R
\name{e_opts_xAxis}
\alias{e_opts_xAxis}
\title{Get x-axis options}
\usage{
e_opts_xAxis(e)
}
\arguments{
\item{e}{\code{htmlwidget/list} An echarts objects or an options list}
}
\value{
\code{list} of options
}
\description{
Get x-axis options
}
\examples{
lapply(list(echarty::ec.init(mtcars), echarts4r::e_chart(mtcars)), e_opts_xAxis)
}
