% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_general.R
\name{e_parallel_default_opts}
\alias{e_parallel_default_opts}
\title{Use the default options for parallelAxis ECharts}
\usage{
e_parallel_default_opts(
  e,
  dims = TRUE,
  grid = list(containLabel = TRUE),
  toolbox = list(orient = "vertical", itemSize = 8, itemGap = 3),
  tooltip = list(trigger = "axis"),
  saveAsImage_filename = list(),
  parallelAxisDefault = TRUE,
  ...
)
}
\arguments{
\item{dims}{\code{(list/logical/NULL)} See \link[echarts4r]{e_dims}}

\item{toolbox}{\code{(list/logical/NULL)} See \link[echarts4r]{e_toolbox}}

\item{tooltip}{\code{(list/logical/NULL)} See \link[echarts4r]{e_tooltip}}

\item{saveAsImage_filename}{\code{(list/logical/NULL)} See \code{saveAsImage_filename}}

\item{...}{Additional \code{echarts4r} named arguments, IE for \link[echarts4r]{e_add_nested} used \verb{add_nested = list([...])} where the list holds the args to pass to the function.}
}
\value{
\code{echart}
}
\description{
Use the default options for parallelAxis ECharts
}
