% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_general.R
\name{e_parallelAxisDefault}
\alias{e_parallelAxisDefault}
\title{Add default options for Parallel Axis}
\usage{
e_parallelAxisDefault(
  e,
  nameLocation = "end",
  nameGap = 15,
  axisLabel = list(formatter = js_num2str()),
  ...
)
}
\arguments{
\item{nameLocation}{\code{chr} \href{https://echarts.apache.org/en/option.html#parallel.parallelAxisDefault.nameLocation}{Link}}

\item{nameGap}{\code{num} \href{https://echarts.apache.org/en/option.html#parallel.parallelAxisDefault.nameGap}{Link}}

\item{axisLabel}{\code{list} \href{https://echarts.apache.org/en/option.html#parallel.parallelAxisDefault.axisLabel}{Link}}
}
\value{
\code{echart}
}
\description{
Add default options for Parallel Axis
}
