% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_axes.R
\name{e_series_axis_data}
\alias{e_series_axis_data}
\title{Extract data for a particular axis from an `axis_data`` tbl}
\usage{
e_series_axis_data(axis_data, axis = "x", flat = TRUE)
}
\arguments{
\item{axis_data}{\code{tbl} Axis data as returned by \code{e_series_data}}

\item{axis}{\code{chr} x or y. z currently not supported.}

\item{flat}{\code{lgl} Whether to flatten or leave as a tbl}
}
\value{
\code{num/tbl} depending on the value to \code{flat}
}
\description{
Extract data for a particular axis from an `axis_data`` tbl
}
\seealso{
e_series_data
}
