% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{e_series_data}
\alias{e_series_data}
\title{Extract the series data}
\usage{
e_series_data(e, axis = "y", as_chr = TRUE)
}
\arguments{
\item{e}{\code{echart}}

\item{axis}{\code{chr} "x" or "y" to indicate data for which axis}

\item{as_chr}{\code{lgl} conver the output to a human legible string}
}
\value{
\code{list}
}
\description{
Extract the series data
}
