% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_axes.R
\name{e_series_object}
\alias{e_series_object}
\title{Retrieve a specific object from each series item}
\usage{
e_series_object(series, ...)
}
\arguments{
\item{series}{\code{list} Echarts series options}

\item{...}{A list of accessors for indexing into the object. Can be
an positive integer, a negative integer (to index from the right),
a string (to index into names), or an accessor function
(except for the assignment variants which only support names and
positions). If the object being indexed is an S4 object,
accessing it by name will return the corresponding slot.

\link[rlang:dyn-dots]{Dynamic dots} are supported. In particular, if
your accessors are stored in a list, you can splice that in with
\verb{!!!}.}
}
\value{
\code{list}
}
\description{
Retrieve a specific object from each series item
}
\seealso{
Other series: 
\code{\link{e_opts_is_xAxis_series}()},
\code{\link{e_series}()}
}
\concept{series}
