% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_legend.R
\name{e_tooltip_legend_dot}
\alias{e_tooltip_legend_dot}
\title{Create an echarts style legend dot}
\usage{
e_tooltip_legend_dot(color)
}
\arguments{
\item{color}{\code{chr} CSS compliant color}
}
\value{
\code{shiny.tag.list}
}
\description{
Create an echarts style legend dot
}
\examples{
e_tooltip_legend_dot("#48df94")
}
