% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_legend.R
\name{e_tooltip_legend_rectangle}
\alias{e_tooltip_legend_rectangle}
\title{Create a rectangle legend item}
\usage{
e_tooltip_legend_rectangle(
  background_color,
  color = "#000",
  opacity = 0.2,
  height = "10px",
  width = "20px",
  display = "inline-block",
  ...
)
}
\arguments{
\item{background_color}{\code{chr} CSS compliant property}

\item{color}{\code{chr} CSS compliant property}

\item{opacity}{\code{chr} CSS compliant property}

\item{height}{\code{chr} CSS compliant property}

\item{width}{\code{chr} CSS compliant property}

\item{display}{\code{chr} CSS compliant property}

\item{...}{\code{chr} Additional CSS properties}
}
\value{
\code{shiny.tag}
}
\description{
Create a rectangle legend item
}
\examples{
e_tooltip_legend_rectangle("#48df94")
}
\seealso{
Other legend: 
\code{\link{e_tooltip_legend_dot}()}
}
\concept{legend}
