% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_axes.R
\name{e_yAxis_width}
\alias{e_yAxis_width}
\title{Calculate a best guess \code{nameGap} value for spacing the ECharts axis titles beyond the axis labels}
\usage{
e_yAxis_width(x, outtype = c("rounded", "abbreviated"), sf = 0)
}
\arguments{
\item{x}{\code{num} All values on the y axis. See \code{e_series_axis_data}}
}
\value{
\code{num}
}
\description{
Calculate a best guess \code{nameGap} value for spacing the ECharts axis titles beyond the axis labels
}
\seealso{
e_y_axis_formatting
}
