% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/e_opts_axes.R
\name{e_year_on_x}
\alias{e_year_on_x}
\title{Is the X axis in years?}
\usage{
e_year_on_x(e)

e_year_on_x(e)
}
\arguments{
\item{e}{\code{echart}}
}
\value{
\code{lgl}

\code{lgl}
}
\description{
Is the X axis in years?

Is the X axis in years?
}
