% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{ec_toolbox}
\alias{ec_toolbox}
\title{Add default toolbox options to echart
Add ability to save as image, restore, zoom and reset zoom to an echart}
\usage{
ec_toolbox(
  e,
  opts = list(orient = "vertical", feature = list(saveAsImage = list(title =
    shiny::getDefaultReactiveDomain()$ns("chart")), restore = list(), dataZoom = list())),
  opts_merge = NULL,
  opts_modify = NULL,
  minimal = FALSE
)
}
\arguments{
\item{opts}{\code{list} All options to be applied. Comes with defaults.}

\item{opts_merge}{\code{list} Options to be added. See \link[purrr]{list_merge}}

\item{opts_modify}{\code{list} Options to be replaced. See \link[purrr]{list_modify}}

\item{minimal}{\code{boolean} whether or not to add minimal toolbox (currently
defined as just export to image by BoR) to a chart}

\item{p}{\code{echarty}}
}
\value{
\code{echarty}
}
\description{
Add default toolbox options to echart
Add ability to save as image, restore, zoom and reset zoom to an echart
}
