% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echartsUtils-package.R
\docType{package}
\name{echartsUtils-package}
\alias{echartsUtils}
\alias{echartsUtils-package}
\title{echartsUtils: Tools for working with echarts}
\description{
Utility functions for working with echarts objects created by echarts4r & echarty
}
\author{
\strong{Maintainer}: Stephen Holsenbeck \email{sholsen@alumni.emory.edu}

}
\keyword{internal}
