% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_helpers.R
\name{use_echartsUtils_deps}
\alias{use_echartsUtils_deps}
\title{Use dependencies for echartsUtils}
\usage{
use_echartsUtils_deps()
}
\value{
\code{shiny.tag.list}
}
\description{
Use dependencies for echartsUtils
}
