% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_general.R, R/options.R
\name{e_has_group}
\alias{e_has_group}
\title{Does the echart have a group}
\usage{
e_has_group(e)

e_has_group(e)
}
\arguments{
\item{e}{\code{echart}}
}
\value{
\code{lgl}

\code{lgl}
}
\description{
Does the echart have a group

Does the echart have a group
}
\seealso{
Other options: 
\code{\link{e_default_opts}()},
\code{\link{e_get_opts}()},
\code{\link{e_has_base_opts}()},
\code{\link{e_has_timeline}()},
\code{\link{e_opts_is_xAxis_series}()},
\code{\link{e_opts_line_default}()},
\code{\link{e_opts_make_stacked_area_series}()},
\code{\link{e_opts_modify}()},
\code{\link{e_opts_update}()},
\code{\link{e_opts_xAxis_type}()},
\code{\link{e_opts_xAxis}()},
\code{\link{e_opts_yAxis}()},
\code{\link{e_series_axis_data}()},
\code{\link{e_series_data}()},
\code{\link{e_series}()},
\code{\link{e_x_axis_formatting}()},
\code{\link{e_yAxis_width}()},
\code{\link{e_y_axis_formatting}()}
}
\concept{options}
