% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_general.R
\name{e_opt_modify}
\alias{e_opt_modify}
\title{Modify an echarts option that may be an unnamed or named list}
\usage{
e_opt_modify(opt, new_opts)
}
\arguments{
\item{opt}{\code{list} An echart option list}

\item{new_opts}{\code{list} options to add}
}
\value{
\code{list} the modified option
}
\description{
Modify an echarts option that may be an unnamed or named list
}
