% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{e_opts_is_xAxis_series}
\alias{e_opts_is_xAxis_series}
\title{Does the echart have xAxis options for each series (multiple unnamed lists with options)}
\usage{
e_opts_is_xAxis_series(e)
}
\arguments{
\item{e}{\code{echart}}
}
\value{
\code{lgl}
}
\description{
Does the echart have xAxis options for each series (multiple unnamed lists with options)
}
