% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_line.R
\name{e_opts_line_default}
\alias{e_opts_line_default}
\title{A list of default options for an echart line}
\usage{
e_opts_line_default(
  .data,
  grid = list(left = "20\%"),
  color = NULL,
  tooltip = list(trigger = "axis", axisPointer = list(type = "cross")),
  legend = list(textStyle = list(fontSize = 10), type = "scroll", bottom = "0\%", opacity
    = 1, padding = list(0, 0, 0, 0), itemGap = 2),
  xAxis = list(type = "category"),
  yAxis = list(type = "value"),
  ...
)
}
\arguments{
\item{.data}{\code{tbl} of Reservoir Data}

\item{grid}{See \href{https://echarts.apache.org/en/option.html#grid}{echarts setOption docs}}

\item{color}{See \href{https://echarts.apache.org/en/option.html#color}{echarts setOption docs}}

\item{tooltip}{See \href{https://echarts.apache.org/en/option.html#tooltip}{echarts setOption docs}}

\item{legend}{See \href{https://echarts.apache.org/en/option.html#legend}{echarts setOption docs}}

\item{xAxis}{See \href{https://echarts.apache.org/en/option.html#xAxis}{echarts setOption docs}}

\item{yAxis}{See \href{https://echarts.apache.org/en/option.html#yAxis}{echarts setOption docs}}

\item{...}{\code{args} Additional parameters}
}
\value{
\code{list}
}
\description{
A list of default options for an echart line
}
\seealso{
Other line: 
\code{\link{e_opts_make_stacked_area_series}()}

Other options: 
\code{\link{e_default_opts}()},
\code{\link{e_get_opts}()},
\code{\link{e_has_base_opts}()},
\code{\link{e_has_group}()},
\code{\link{e_has_timeline}()},
\code{\link{e_opts_is_xAxis_series}()},
\code{\link{e_opts_make_stacked_area_series}()},
\code{\link{e_opts_modify}()},
\code{\link{e_opts_update}()},
\code{\link{e_opts_xAxis_type}()},
\code{\link{e_opts_xAxis}()},
\code{\link{e_opts_yAxis}()},
\code{\link{e_series_axis_data}()},
\code{\link{e_series_data}()},
\code{\link{e_series}()},
\code{\link{e_x_axis_formatting}()},
\code{\link{e_yAxis_width}()},
\code{\link{e_y_axis_formatting}()}
}
\concept{line}
\concept{options}
