% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_axes.R
\name{e_opts_xAxis}
\alias{e_opts_xAxis}
\title{Get x-axis options}
\usage{
e_opts_xAxis(e)
}
\arguments{
\item{e}{\code{htmlwidget/list} An echarts objects or an options list}
}
\value{
\code{list} of options
}
\description{
Get x-axis options
}
\examples{
lapply(list(echarty::ec.init(mtcars), echarts4r::e_chart(mtcars)), e_opts_xAxis)
}
\seealso{
Other options: 
\code{\link{e_default_opts}()},
\code{\link{e_get_opts}()},
\code{\link{e_has_base_opts}()},
\code{\link{e_has_group}()},
\code{\link{e_has_timeline}()},
\code{\link{e_opts_is_xAxis_series}()},
\code{\link{e_opts_line_default}()},
\code{\link{e_opts_make_stacked_area_series}()},
\code{\link{e_opts_modify}()},
\code{\link{e_opts_parallelAxisDefault}()},
\code{\link{e_opts_parallel}()},
\code{\link{e_opts_update}()},
\code{\link{e_opts_xAxis_type}()},
\code{\link{e_opts_yAxis}()},
\code{\link{e_parallel_default_opts}()},
\code{\link{e_series_axis_data}()},
\code{\link{e_series_data}()},
\code{\link{e_series}()},
\code{\link{e_x_axis_formatting}()},
\code{\link{e_yAxis_width}()},
\code{\link{e_y_axis_formatting}()}
}
\concept{options}
