% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_general.R
\name{e_parallel_default_opts}
\alias{e_parallel_default_opts}
\title{Use the default options for parallelAxis ECharts}
\usage{
e_parallel_default_opts(
  e,
  dims = TRUE,
  grid = list(containLabel = TRUE),
  toolbox = list(orient = "vertical", itemSize = 8, itemGap = 3),
  tooltip = list(trigger = "axis"),
  saveAsImage_filename = list(),
  parallelAxisDefault = TRUE,
  ...
)
}
\arguments{
\item{dims}{\code{(list/logical/NULL)} See \link[echarts4r]{e_dims}}

\item{toolbox}{\code{(list/logical/NULL)} See \link[echarts4r]{e_toolbox}}

\item{tooltip}{\code{(list/logical/NULL)} See \link[echarts4r]{e_tooltip}}

\item{saveAsImage_filename}{\code{(list/logical/NULL)} See \code{saveAsImage_filename}}

\item{...}{Additional \code{echarts4r} named arguments, IE for \link[echarts4r]{e_add_nested} used \verb{add_nested = list([...])} where the list holds the args to pass to the function.}
}
\value{
\code{echart}
}
\description{
Use the default options for parallelAxis ECharts
}
\seealso{
Other options: 
\code{\link{e_default_opts}()},
\code{\link{e_get_opts}()},
\code{\link{e_has_base_opts}()},
\code{\link{e_has_group}()},
\code{\link{e_has_timeline}()},
\code{\link{e_opts_is_xAxis_series}()},
\code{\link{e_opts_line_default}()},
\code{\link{e_opts_make_stacked_area_series}()},
\code{\link{e_opts_modify}()},
\code{\link{e_opts_parallelAxisDefault}()},
\code{\link{e_opts_parallel}()},
\code{\link{e_opts_update}()},
\code{\link{e_opts_xAxis_type}()},
\code{\link{e_opts_xAxis}()},
\code{\link{e_opts_yAxis}()},
\code{\link{e_series_axis_data}()},
\code{\link{e_series_data}()},
\code{\link{e_series}()},
\code{\link{e_x_axis_formatting}()},
\code{\link{e_yAxis_width}()},
\code{\link{e_y_axis_formatting}()}
}
\concept{options}
