% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{e_saveAsImage_filename}
\alias{e_saveAsImage_filename}
\title{saveAsImage_filename}
\usage{
e_saveAsImage_filename(e, watermark = FALSE, name = "chart_download.png")
}
\value{
A modified echarts obj
}
\description{
A utils function used in echarts module building chains. The
function takes an echarts object in, as well as a name and adds the name to
the saveImageAs toolbox feature.
}
