% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_axes.R
\name{e_series_axis_data}
\alias{e_series_axis_data}
\title{Extract data for a particular axis from an `axis_data`` tbl}
\usage{
e_series_axis_data(axis_data, axis = "x", flat = TRUE)
}
\arguments{
\item{axis_data}{\code{tbl} Axis data as returned by \code{e_series_data}}

\item{axis}{\code{chr} x or y. z currently not supported.}

\item{flat}{\code{lgl} Whether to flatten or leave as a tbl}
}
\value{
\code{num/tbl} depending on the value to \code{flat}
}
\description{
Extract data for a particular axis from an `axis_data`` tbl
}
\seealso{
e_series_data

Other options: 
\code{\link{e_default_opts}()},
\code{\link{e_get_opts}()},
\code{\link{e_has_base_opts}()},
\code{\link{e_has_group}()},
\code{\link{e_has_timeline}()},
\code{\link{e_opts_is_xAxis_series}()},
\code{\link{e_opts_line_default}()},
\code{\link{e_opts_make_stacked_area_series}()},
\code{\link{e_opts_modify}()},
\code{\link{e_opts_update}()},
\code{\link{e_opts_xAxis_type}()},
\code{\link{e_opts_xAxis}()},
\code{\link{e_opts_yAxis}()},
\code{\link{e_series_data}()},
\code{\link{e_series}()},
\code{\link{e_x_axis_formatting}()},
\code{\link{e_yAxis_width}()},
\code{\link{e_y_axis_formatting}()}

Other data: 
\code{\link{e_series_data}()},
\code{\link{e_year_on_x}()}
}
\concept{data}
\concept{options}
