% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_axes.R
\name{e_series_data}
\alias{e_series_data}
\title{Extract the series data}
\usage{
e_series_data(e)
}
\arguments{
\item{e}{\code{echart}}

\item{axis}{\code{chr} "x" or "y" to indicate data for which axis}

\item{as_chr}{\code{lgl} conver the output to a human legible string}
}
\value{
\code{list}
}
\description{
Extract the series data
}
\seealso{
Other options: 
\code{\link{e_default_opts}()},
\code{\link{e_get_opts}()},
\code{\link{e_has_base_opts}()},
\code{\link{e_has_group}()},
\code{\link{e_has_timeline}()},
\code{\link{e_opts_is_xAxis_series}()},
\code{\link{e_opts_line_default}()},
\code{\link{e_opts_make_stacked_area_series}()},
\code{\link{e_opts_modify}()},
\code{\link{e_opts_update}()},
\code{\link{e_opts_xAxis_type}()},
\code{\link{e_opts_xAxis}()},
\code{\link{e_opts_yAxis}()},
\code{\link{e_series_axis_data}()},
\code{\link{e_series}()},
\code{\link{e_x_axis_formatting}()},
\code{\link{e_yAxis_width}()},
\code{\link{e_y_axis_formatting}()}

Other data: 
\code{\link{e_series_axis_data}()},
\code{\link{e_year_on_x}()}
}
\concept{data}
\concept{options}
