% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_formatting.R
\name{e_x_axis_formatting}
\alias{e_x_axis_formatting}
\title{e_x_axis_year_formatting}
\usage{
e_x_axis_formatting(
  e,
  scale = TRUE,
  type = e_opts_xAxis_type(e) \%||\% "value",
  min = purrr::when(e_year_on_x(e), isTRUE(.) ~ 1940, ~NULL),
  max = purrr::when(e_year_on_x(e), isTRUE(.) ~ 2060, ~NULL),
  name = purrr::when(e_year_on_x(e), isTRUE(.) ~ "Year", ~NULL),
  nameLocation = "middle",
  nameTextStyle = list(padding = rep(0, 4), fontFamily = "rennerbook"),
  nameGap = 30,
  rotate = 45,
  fontFamily = "rennerbook",
  ...
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[echarts4r:axis]{echarts4r::e_x_axis}}
  \describe{
    \item{\code{e}}{An \code{echarts4r} object as returned by \code{\link[echarts4r]{e_charts}} or
a proxy as returned by \code{\link[echarts4r]{echarts4rProxy}}.}
    \item{\code{serie}}{Column name of serie to range the axis. If used the range of the serie is used as,
\code{min} an \code{max}.}
    \item{\code{index}}{Index of axis to customise.}
    \item{\code{formatter}}{An axis formatter as returned by \code{\link[echarts4r]{e_axis_formatter}}.}
    \item{\code{margin}}{Margin to apply to \code{serie}: \eqn{min = serie - margin} and
\eqn{max = serie + margin}}
  }}

\item{e_ctxt}{\code{(character)} The context in which the e_chart is being rendered}
}
\value{
A list of parameters for appropriate x-axis formatting for years on echarts figures
}
\description{
generates a list that configures the min/max and correct axis formatting
for echarts figs
}
