% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_axes.R
\name{e_y_axis_formatting}
\alias{e_y_axis_formatting}
\title{e_y_axis_year_formatting}
\usage{
e_y_axis_formatting(
  e,
  scale = TRUE,
  type = "value",
  formatter = echartsUtils::js_num2str(),
  name = NULL,
  nameLocation = "middle",
  nameTextStyle = list(padding = c(0, 0, 0, 0)),
  nameGap = "auto",
  mo = NULL,
  axisPointer = e_opts_axisPointer(),
  ...
)
}
\arguments{
\item{name}{\code{chr} Axis label}

\item{nameLocation}{\code{chr} Where to position the axis label. \strong{Default 'middle'}}

\item{nameTextStyle}{\code{list} With named parameters. See \href{https://echarts.apache.org/en/option.html#yAxis.nameTextStyle}{Echarts yAxis.nameTextStyle}}

\item{nameGap}{\code{num/chr} \strong{Default: 'auto'} uses custom heuristics to determine a best guess. See \link{axis_width}. Otherwise a numeric pixel value}

\item{mo}{\code{int} of the magnitude of the units. 9 = Billions, 6 = Millions etc}

\item{...}{
  Arguments passed on to \code{\link[echarts4r:axis]{echarts4r::e_y_axis}}
  \describe{
    \item{\code{e}}{An \code{echarts4r} object as returned by \code{\link[echarts4r]{e_charts}} or
a proxy as returned by \code{\link[echarts4r]{echarts4rProxy}}.}
    \item{\code{serie}}{Column name of serie to range the axis. If used the range of the serie is used as,
\code{min} an \code{max}.}
    \item{\code{index}}{Index of axis to customise.}
    \item{\code{formatter}}{An axis formatter as returned by \code{\link[echarts4r]{e_axis_formatter}}.}
    \item{\code{margin}}{Margin to apply to \code{serie}: \eqn{min = serie - margin} and
\eqn{max = serie + margin}}
  }}

\item{fontFamily}{\code{chr} CSS font-family}
}
\value{
A list of parameters for appropriate y-axis formatting for years on echarts figures
}
\description{
generates a list that configures the correct y axis formatting for echarts figs
}
\seealso{
echarts4r::e_y_axis

Other options: 
\code{\link{e_default_opts}()},
\code{\link{e_get_opts}()},
\code{\link{e_has_base_opts}()},
\code{\link{e_has_group}()},
\code{\link{e_has_timeline}()},
\code{\link{e_opts_is_xAxis_series}()},
\code{\link{e_opts_line_default}()},
\code{\link{e_opts_make_stacked_area_series}()},
\code{\link{e_opts_modify}()},
\code{\link{e_opts_parallelAxisDefault}()},
\code{\link{e_opts_parallel}()},
\code{\link{e_opts_update}()},
\code{\link{e_opts_xAxis_type}()},
\code{\link{e_opts_xAxis}()},
\code{\link{e_opts_yAxis}()},
\code{\link{e_parallel_default_opts}()},
\code{\link{e_series_axis_data}()},
\code{\link{e_series_data}()},
\code{\link{e_series}()},
\code{\link{e_x_axis_formatting}()},
\code{\link{e_yAxis_width}()}
}
\concept{options}
