% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_opts_axes.R
\name{e_yAxis_width}
\alias{e_yAxis_width}
\title{Calculate a best guess \code{nameGap} value for spacing the ECharts axis titles beyond the axis labels}
\usage{
e_yAxis_width(x, outtype = c("rounded", "abbreviated"), sf = 0)
}
\arguments{
\item{x}{\code{num} All values on the y axis. See \code{e_series_axis_data}}
}
\value{
\code{num}
}
\description{
Calculate a best guess \code{nameGap} value for spacing the ECharts axis titles beyond the axis labels
}
\seealso{
e_y_axis_formatting

Other options: 
\code{\link{e_default_opts}()},
\code{\link{e_get_opts}()},
\code{\link{e_has_base_opts}()},
\code{\link{e_has_group}()},
\code{\link{e_has_timeline}()},
\code{\link{e_opts_is_xAxis_series}()},
\code{\link{e_opts_line_default}()},
\code{\link{e_opts_make_stacked_area_series}()},
\code{\link{e_opts_modify}()},
\code{\link{e_opts_update}()},
\code{\link{e_opts_xAxis_type}()},
\code{\link{e_opts_xAxis}()},
\code{\link{e_opts_yAxis}()},
\code{\link{e_series_axis_data}()},
\code{\link{e_series_data}()},
\code{\link{e_series}()},
\code{\link{e_x_axis_formatting}()},
\code{\link{e_y_axis_formatting}()}
}
\concept{options}
